feat: add preview url, fix files

+++ webui.py
@@ -31,6 +31,7 @@ cfg = {
     "DRAW_TIMESTAMP": config.DRAW_TIMESTAMP,
     "FONT": config.FONT,
     "TEXT_STYLE": config.TEXT_STYLE,
+    "PREVIEW_URL": "" if not hasattr(config, "PREVIEW_URL") else config.PREVIEW_URL,
 }
 
 if cfg.get("URL") and not cfg.get("RECEIVER_URL"):
@@ -83,6 +84,7 @@ def update_settings():
         ),
         "FONT": form_data.get("FONT", default_config.FONT),
         "TEXT_STYLE": form_data.get("TEXT_STYLE", default_config.TEXT_STYLE),
+        "PREVIEW_URL": form_data.get("PREVIEW_URL", ""),
     }
 
     if cfg.get("TOKEN"):
@@ -96,27 +98,28 @@ def update_settings():
 
     if "@" in cam_ip:
         cam_ip = f"{cam_ip.split('://')[0]}://{cam_ip.split('@')[1]}"
-    cam_ip = cam_ip.split("://")[1].split(":")[0]
 
     # Update settings (write a new config_override.py file)
     with open("config_override.py", "w") as f:
         for k, v in cfg.items():
             f.write(f"{k} = {repr(v)}\n")
 
+    cam_ip = cam_ip.split("://")[-1].split(":")[0]
     # Update the camera url in nginx
-    os.system("rm /etc/nginx/sites-enabled/myproxy.conf")
-    with open("/etc/nginx/sites-enabled/myproxy.conf", "w+") as file:
-        file.write(
-            f"server {{ listen 8080; location / {{ proxy_pass https://{cam_ip}:443; }} }}"
-        )
+    if os.path.exists("/etc/nginx/sites-enabled"):
+        os.system("rm /etc/nginx/sites-enabled/myproxy.conf")
+        with open("/etc/nginx/sites-enabled/myproxy.conf", "w+") as file:
+            file.write(
+                f"server {{ listen 8080; location / {{ proxy_pass https://{cam_ip}:443; }} }}"
+            )
 
-    router_ip = ".".join(cam_ip.split(".")[0:3]) + ".1"
-    os.system("rm /etc/nginx/sites-enabled/router.conf")
-    with open("/etc/nginx/sites-enabled/router.conf", "w+") as file:
-        file.write(
-            f"server {{ listen 8081; location / {{ proxy_pass http://{router_ip}/; proxy_set_header Referer http://{router_ip}/; }} }}"
-        )
-    os.system("service nginx restart")
+        router_ip = ".".join(cam_ip.split(".")[0:3]) + ".1"
+        os.system("rm /etc/nginx/sites-enabled/router.conf")
+        with open("/etc/nginx/sites-enabled/router.conf", "w+") as file:
+            file.write(
+                f"server {{ listen 8081; location / {{ proxy_pass http://{router_ip}/; proxy_set_header Referer http://{router_ip}/; }} }}"
+            )
+        os.system("service nginx restart")
 
     return render_template("config.html", config=cfg)
